still trying to get this damn migration to go through

+++ db/migrate_cards/20180105134604_created_at_for_answers.rb
@@ -3,10 +3,10 @@
 class CreatedAtForAnswers < Card::Migration
   def up
     Answer.where(created_at: nil).each do |a|
+      next unless a.card
       if a.card.type_id == Card::MetricValueID
         a.update_attributes! created_at: a.card.created_at
       else
-        a.card.delete!
         a.destroy!
       end
     end
